redis缓存的配置和使用

+++ e3_content/e3_content_service/src/main/java/com/e3mall/content/service/impl/ContentServiceImpl.java
@@ -10,11 +10,16 @@ import java.util.ArrayList;
 import java.util.Date;
 import java.util.List;
 
+import org.apache.commons.lang3.StringUtils;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Service;
 
+import com.alibaba.dubbo.common.utils.Log;
+import com.e3mall.common.jedis.JedisClient;
 import com.e3mall.common.pojo.E3Result;
 import com.e3mall.common.pojo.EasyUIDataGridResult;
+import com.e3mall.common.utils.JsonUtils;
 import com.e3mall.content.service.ContentService;
 import com.e3mall.mapper.TbContentMapper;
 import com.e3mall.pojo.TbContent;
@@ -34,6 +39,11 @@ public class ContentServiceImpl implements ContentService {
 	
 	@Autowired
 	private TbContentMapper contentMapper;
+	@Autowired
+	private JedisClient jedisClient;
+	
+	@Value("${CONTENT_LIST}")
+	private String CONTENT_LIST;
 
 	/* 
 	 * @see com.e3mall.content.service.ContentService#getContentList(java.lang.Long, java.lang.Integer, java.lang.Integer) 
@@ -67,6 +77,8 @@ public class ContentServiceImpl implements ContentService {
 			tbContent.setUpdated(new Date());
 			//执行保存操作
 			contentMapper.insert(tbContent);
+			//缓存同步,删除缓存中对应的数据
+			jedisClient.hdel(CONTENT_LIST, tbContent.getCategoryId().toString());
 		} catch (Exception e) {
 			e.printStackTrace();
 			return E3Result.build(500, "发生系统错误!");
@@ -84,6 +96,8 @@ public class ContentServiceImpl implements ContentService {
 			tbContent.setUpdated(new Date());
 			//执行业务方法
 			contentMapper.updateByPrimaryKey(tbContent);
+			//缓存同步,删除缓存中对应的数据
+			jedisClient.hdel(CONTENT_LIST, tbContent.getCategoryId().toString());
 		} catch (Exception e) {
 			e.printStackTrace();
 			return E3Result.build(500, "系统错误");
@@ -97,7 +111,18 @@ public class ContentServiceImpl implements ContentService {
 	@Override
 	public E3Result deleteBatchContentByIds(ArrayList<String> ids) {
 		try {
-			contentMapper.deleteBatchById(ids);
+			if (!ids.isEmpty()) {
+				//获取内容id
+				Long id = new Long(ids.get(0));
+				//查询出对应cid
+				TbContent tbContent = contentMapper.selectByPrimaryKey(id);
+				
+				//执行批量删除业务
+				contentMapper.deleteBatchById(ids);		
+				
+				//缓存同步,删除缓存中对应的数据
+				jedisClient.hdel(CONTENT_LIST, tbContent.getCategoryId().toString());				
+			}
 		} catch (Exception e) {
 			e.printStackTrace();
 			return E3Result.build(500, "系统错误");
@@ -111,13 +136,25 @@ public class ContentServiceImpl implements ContentService {
 	@Override
 	public List<TbContent> getContentListById(Long cid) {
 		List<TbContent> list = null;
-		try {			
+		try {		
+			//查询缓存
+			String json = jedisClient.hget(CONTENT_LIST, cid + "");
+			//判断json是否为空
+			if (StringUtils.isNotBlank(json)) {
+				//把json转换成list
+				list = JsonUtils.jsonToList(json, TbContent.class);
+				return list;
+			}
+			
 			TbContentExample example = new TbContentExample();
 			Criteria criteria = example.createCriteria();
 			//设置查询条件
 			criteria.andCategoryIdEqualTo(cid);
 			//执行查询
 			list = contentMapper.selectByExampleWithBLOBs(example);
+			
+			//向集合中添加缓存
+			jedisClient.hset(CONTENT_LIST, cid + "", JsonUtils.objectToJson(list));
 		} catch (Exception e) {
 			e.printStackTrace();
 			return list;
